fix(catalog): address sixth round of PR review feedback
- Reject catalog.acceleration for any provider other than pg at
  config-build time instead of silently no-op'ing it (Copilot):
  acceleration is only implemented for the PostgreSQL catalog
  connector, and every other provider's connector ignored the field
  entirely.
- Extract the CDC-acceleration-support PostgreSQL queries (list_schemas,
  list_tables, primary_key_columns, check_cdc_prerequisites) into a new
  connector-postgres-common crate, per Phillip's review comment, so the
  same logic isn't duplicated across postgres-related crates and so
  runtime never needs a direct dependency on a connector-* crate.
  data_components::postgres::provider now depends on and re-exports
  these functions, so existing call sites are unaffected.
