+++ crates/data-connectors/connector-postgres-common/Cargo.toml
@@ -1,0 +1,32 @@
+##
+## Copyright 2024-2026 The Spice.ai OSS Authors
+##
+## Licensed under the Apache License, Version 2.0 (the "License");
+## you may not use this file except in compliance with the License.
+## You may obtain a copy of the License at
+##
+##      https://www.apache.org/licenses/LICENSE-2.0
+##
+## Unless required by applicable law or agreed to in writing, software
+## distributed under the License is distributed on an "AS IS" BASIS,
+## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+## See the License for the specific language governing permissions and
+## limitations under the License.
+##
+
+[package]
+name = "connector-postgres-common"
+edition.workspace = true
+version.workspace = true
+rust-version.workspace = true
+license-file.workspace = true
+description = "Shared PostgreSQL catalog/schema introspection utilities used by multiple Spice.ai PostgreSQL-related crates"
+
+[dependencies]
+datafusion-common.workspace = true
+datafusion-table-providers = { workspace = true, features = ["postgres"] }
+snafu.workspace = true
+tokio-postgres.workspace = true
+
+[lints]
+workspace = true
